Re-implement like loading
FB API doesn't work with offset greater than 1000 so we'll need to use the paging to load all likes instead. Create a LikeService to handle loading pages until the final one is reached. This also allows using the data range (although it makes the loading bar unreliable).

FbApi in UserService is simplified and setting parameters is left to the caller.